+++ app/Http/Controllers/RoleController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
 
@@ -14,9 +15,12 @@ class RoleController extends Controller
     public function index()
     {
         //
+        if(!Gate::allows('role_list')){
+            abort(401);
+        }
         $roles = Role::with('permissions')->get();
         // dd($roles);
-        return view('roles.index',compact('roles'));
+        return view('roles.index', compact('roles'));
     }
 
     /**
@@ -25,8 +29,11 @@ class RoleController extends Controller
     public function create()
     {
         //
+        if(!Gate::allows('role_create')){
+            abort(401);
+        }
         $permissions = Permission::all();
-        return view('roles.create',compact('permissions'));
+        return view('roles.create', compact('permissions'));
     }
 
     /**
@@ -36,6 +43,9 @@ class RoleController extends Controller
     {
         //
         // dd($request->all());
+        if(!Gate::allows('role_store')){
+            abort(401);
+        }
         $request->validate([
             'role' => ['required'],
             'permissions' => ['required']
@@ -53,7 +63,7 @@ class RoleController extends Controller
     public function show(string $id)
     {
         //
-        
+
     }
 
     /**
@@ -62,10 +72,13 @@ class RoleController extends Controller
     public function edit(string $id)
     {
         //
+        if(!Gate::allows('role_edit')){
+            abort(401);
+        }
         $permissions = Permission::all();
-        $role = Role::with('permissions')->where('id',$id)->first();
+        $role = Role::with('permissions')->where('id', $id)->first();
         // dd($permissions);
-        return view('roles.edit',compact('role','permissions'));
+        return view('roles.edit', compact('role', 'permissions'));
     }
 
     /**
@@ -74,21 +87,22 @@ class RoleController extends Controller
     public function update(Request $request, string $id)
     {
         //
+        if(!Gate::allows('role_update')){
+            abort(401);
+        }
         $request->validate([
             'role' => ['required'],
             'permissions' => ['required']
         ]);
-      $role = Role::where('id',$id)->first();
-      // Update the role's name
-    $role->name = $request->input('role');
-    $role->save();
+        $role = Role::where('id', $id)->first();
+        // Update the role's name
+        $role->name = $request->input('role');
+        $role->save();
 
-    // Sync the role's permissions
-    $permissions = $request->input('permissions', []);
-    $role->syncPermissions($permissions);
-    return redirect()->route('roles.index');
-
-        
+        // Sync the role's permissions
+        $permissions = $request->input('permissions', []);
+        $role->syncPermissions($permissions);
+        return redirect()->route('roles.index');
     }
 
     /**
@@ -97,9 +111,10 @@ class RoleController extends Controller
     public function destroy(string $id)
     {
         //
-        Role::where('id',$id)->delete();
+        if(!Gate::allows('role_destory')){
+            abort(401);
+        }
+        Role::where('id', $id)->delete();
         return redirect()->route('roles.index');
-
-
     }
 }
